+++ src/actions/boards.js
@@ -1,0 +1,16 @@
+import * as actionTypes from '../utils/actionTypes'
+
+export const openSelectedBoard = (boardId) => ({
+  type: actionTypes.openSelectedBoard,
+  boardId,
+})
+
+export const backToBoardList = () => ({
+  type: actionTypes.backToBoardList,
+})
+
+export const registerNewColumnToBoard = (boardId, newColumnId) => ({
+  type: actionTypes.registerNewColumnToBoard,
+  boardId,
+  newColumnId,
+})
