+++ src/container/TaskChuteContainer.js
@@ -5,7 +5,7 @@ import PropTypes from 'prop-types';
 
 import ActiveBoard from '../components/ActiveBoard';
 import BoardList from '../components/BoardList';
-import * as actions from '../actions';
+import actions from '../actions';
 import '../css/main.css';
 
 class TaskChuteContainer extends React.Component {
